# openapi.yaml
openapi: 3.1.0
info:
  title: DeenBridge Backend API
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/Deen-Bridge/dnb-backend/blob/main/LICENSE
  description: |
    REST API for DeenBridge: authentication and sessions, courses, the book library,
    community spaces, reels, search, Stellar USDC payments, refunds and disputes,
    educator payouts, media upload signing and the SEP-1 metadata document.

    Provenance of this document:
      - Paths, methods and auth requirements are derived from app.js and src/routes/**,
        which are the authoritative mounting and middleware sources.
      - Component schemas are derived from the Mongoose schemas in src/models/**.
      - Per-endpoint request bodies are described from route middleware and the model
        fields they write. Where a controller accepts more than the model exposes, the
        body is intentionally permissive rather than invented.

    Known response inconsistencies (see issue about response standardization):
      - Most handlers return an envelope of the form { success, message, data }.
      - Some list handlers return a bare array with no envelope.
      - POST /api/email returns { success, otp, message } on success but a bare
        { message } on validation failure.
      - GET /admin/jobs returns HTML when the Accept header prefers HTML, otherwise
        { success, jobs }.
      - GET /metrics returns Prometheus text, and the SEP-1 document returns text/toml.
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by POST /api/auth/login or POST /api/auth/refresh.
    jobsToken:
      type: http
      scheme: bearer
      description: Static token from the JOBS_DASHBOARD_TOKEN environment variable.
...
  responses:
    Ok:
      description: Success envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessResponse" }
...
    BadRequest:
      description: Missing or invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed or expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted for this resource or role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist, or the feature is disabled by configuration
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ObjectId:
      type: string
      description: MongoDB ObjectId as a 24 character hex string
      pattern: "^[a-fA-F0-9]{24}$"
    Error:
      type: object
      description: Standard error shape produced by the central error handler
      properties:
        success: { type: boolean, examples: [false] }
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - RATE_LIMITED
                - PAYLOAD_TOO_LARGE
                - REQUEST_TIMEOUT
                - NETWORK_UNAVAILABLE
                - PAYMENT_FAILED
                - PAYMENT_REJECTED
                - WALLET_MISMATCH
                - WALLET_NOT_CONNECTED
                - ITEM_NOT_FOUND
                - ALREADY_PURCHASED
                - INSUFFICIENT_BALANCE
                - STALE_XDR
                - INVALID_XDR
                - SPONSORSHIP_REJECTED
                - INTERNAL_ERROR
                - DATABASE_ERROR
                - EXTERNAL_SERVICE_ERROR
            message: { type: string, description: Human-readable error message }
            details:
              type: array
              description: Optional field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        stack: { type: string, description: Present only in development }
        reqId: { type: string, description: Request correlation ID }
    SuccessResponse:
      type: object
      description: Standard success envelope. Some list handlers return a bare array instead.
      properties:
        success: { type: boolean, examples: [true] }
        message: { type: string }
        data: { description: Payload, shape depends on the endpoint }
    StellarPublicKey:
      type: string
      pattern: "^G[A-Z0-9]{55}$"
      description: Stellar account public key
...
    User:
      type: object
      description: Account. The password field is never returned by the API.
      properties:
        _id: { $ref: "#/components/schemas/ObjectId" }
        name: { type: string }
        email: { type: string, format: email }
        avatar: { type: string, format: uri }
        gender: { type: string, enum: [male, female] }
        age: { type: integer, minimum: 2, maximum: 120 }
        country: { type: string }
        language: { type: string }
        interests:
          type: array
          items: { type: string }
        bio: { type: string, maxLength: 500 }
        role: { type: string, enum: [student, tutor, mentor, admin, arbiter], default: student }
        isActive: { type: boolean, default: true }
        lastLogin: { type: string, format: date-time }
        following:
          type: array
          items: { $ref: "#/components/schemas/ObjectId" }
        followers:
          type: array
          items: { $ref: "#/components/schemas/ObjectId" }
        purchasedBooks:
          type: array
          items:
            type: object
            properties:
              bookId: { $ref: "#/components/schemas/ObjectId" }
              purchaseDate: { type: string, format: date-time }
        purchasedCourses:
          type: array
          items:
            type: object
            properties:
              courseId: { $ref: "#/components/schemas/ObjectId" }
              purchaseDate: { type: string, format: date-time }
        bookmarkedBooks:
          type: array
          items: { $ref: "#/components/schemas/ObjectId" }
        bookmarkedCourses:
          type: array
          items: { $ref: "#/components/schemas/ObjectId" }
        stellarWallet:
          type: object
          properties:
            publicKey: { $ref: "#/components/schemas/StellarPublicKey" }
            connectedAt: { type: string, format: date-time }
            network: { type: string, enum: [testnet, mainnet], default: testnet }
        stat:
          type: object
          properties:
            coursesEnrolled: { type: integer, default: 0 }
            booksRead: { type: integer, default: 0 }
            totalUptime: { type: integer, default: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [name, email]
...
paths:
...
  /api/users/completion:
    get:
      tags: [Users]
      summary: Get authenticated user's profile completion metrics and suggestions
      description: Returns completion percentage, weighted field breakdown, and prioritized improvement suggestions.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/users/recommendations:
    get:
      tags: [Users]
      summary: Personalised recommendations for the caller
      description: Cached per user. Declared before the dynamic id route so it is not shadowed.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
      x-envelope-unverified: true
  /api/users/update/{id}:
    put:
      tags: [Users]
      summary: Update a profile, optionally replacing the avatar
      description: Allowed only for the profile owner or an admin. Invalidates the user cache.
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar: { type: string, format: binary }
                name: { type: string }
                bio: { type: string, maxLength: 500 }
                gender: { type: string, enum: [male, female] }
                age: { type: integer, minimum: 2, maximum: 120 }
                country: { type: string }
                language: { type: string }
                interests: { type: string, description: Repeated field or comma separated list }
          application/json:
            schema: { $ref: "#/components/schemas/User" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Users]
      summary: Delete a user
      description: Allowed only for the account owner or an admin.
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/follow/{id}:
    post:
      tags: [Users]
      summary: Follow a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/unfollow/{id}:
    delete:
      tags: [Users]
      summary: Unfollow a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/{id}/followers:
    get:
      tags: [Users]
      summary: List followers
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
      x-envelope-unverified: true
  /api/users/{id}/following:
    get:
      tags: [Users]
      summary: List accounts a user follows
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
      x-envelope-unverified: true
  /api/users/{id}/followers/count:
    get:
      tags: [Users]
      summary: Follower count
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}/following/count:
    get:
      tags: [Users]
      summary: Following count
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}/check-following:
    get:
      tags: [Users]
      summary: Whether the caller follows this user
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}/stats:
    get:
      tags: [Users]
      summary: Aggregate stats for a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...